+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Kelda Change Log
 
 Up Next
 -------------
-- Only allow object literal arguments to the Container constructor.
+- Only allow object literal arguments to the Container and Infrastructure
+constructors.
 
 Release 0.10.0
 -------------